+++ ktlint-core/src/main/kotlin/com/pinterest/ktlint/core/RuleProvider.kt
@@ -2,7 +2,7 @@ package com.pinterest.ktlint.core
 
 /**
  * Provides a [Rule] instance. Important: to ensure that a [Rule] can keep internal state and that processing of files
- * is thread-safe, a *new* instance should be provided on each call of [createNewRuleInstance]. A custom [RuleProvider]
+ * is thread-safe, a *new* instance should be provided on each call of [getRuleInstance]. A custom [RuleProvider]
  * does not need to take care of clearing internal state as KtLint calls this method any time the [Rule] instance has
  * been used for processing a file and as of that might have an internal state set.
  */
@@ -13,5 +13,5 @@ public class RuleProvider(
      */
     private val provider: () -> Rule,
 ) {
-    public fun createNewRuleInstance(): Rule = provider()
+    public fun getRuleInstance(): Rule = provider()
 }