+++ ktlint-rule-engine-core/src/main/kotlin/com/pinterest/ktlint/rule/engine/core/api/RuleProvider.kt
@@ -4,18 +4,60 @@ import com.pinterest.ktlint.rule.engine.core.api.editorconfig.EditorConfigProper
 import org.ec4j.core.model.PropertyType
 
 /**
- * Provides a [Rule] instance. Important: to ensure that a [Rule] can keep internal state and that processing of files
- * is thread-safe, a *new* instance should be provided on each call of [createNewRuleInstance]. A custom [RuleProvider]
- * does not need to take care of clearing internal state as KtLint calls this method any time the [Rule] instance has
- * been used for processing a file and as of that might have an internal state set.
+ * Provides a [Rule] instance. Important: to ensure that a [Rule] can keep internal state and that processing of files is thread-safe, a
+ * *new* instance should be provided on each call of the [provider] function.
  */
-public class RuleProvider(
+public class RuleProvider private constructor(
     /**
      * Lambda which creates a new instance of the rule.
      */
     private val provider: () -> Rule,
+
+    /**
+     * The rule id of the [Rule] created by the provider.
+     */
+    public val ruleId: RuleId,
+
+    /**
+     * Flag whether the [Rule] created by the provider has to run as late as possible.
+     */
+    public val runAsLateAsPossible: Boolean,
+
+    /**
+     * The list of rules which have to run before the [Rule] created by the provider can be run.
+     */
+    public val runAfterRules: List<Rule.VisitorModifier.RunAfterRule>,
 ) {
-    public fun createNewRuleInstance(): Rule = provider()
+    /**
+     * Creates a new [Rule] instance.
+     */
+    public fun createNewRuleInstance(): Rule {
+        return provider()
+    }
+
+    /**
+     * Lambda which creates a new instance of the [Rule]. Important: to ensure that a [Rule] can keep internal state and that processing of
+     * files is thread-safe, a *new* instance should be provided on each call of the [provider] function.
+     */
+    public companion object {
+        // Note that the KDOC is placed on the companion object to make it actually visually when the RuleProvider identifier is being
+        // hovered in IntelliJ IDEA
+        public operator fun invoke(provider: () -> Rule): RuleProvider =
+            provider()
+                .let { rule ->
+                    RuleProvider(
+                        provider = provider,
+                        ruleId = rule.ruleId,
+                        runAsLateAsPossible = rule
+                            .visitorModifiers
+                            .filterIsInstance<Rule.VisitorModifier.RunAsLateAsPossible>()
+                            .any(),
+                        runAfterRules = rule
+                            .visitorModifiers
+                            .filterIsInstance<Rule.VisitorModifier.RunAfterRule>(),
+                    )
+                }
+    }
 }
 
 /**